+++ src/devices/xiaomi.ts
@@ -77,22 +77,6 @@ const preventReset: OnEvent = async (type, data, device) => {
     await device.getEndpoint(1).write('genBasic', payload, {manufacturerCode});
 };
 
-const daysLookup = {
-    0x7f: 'everyday',
-    0x1f: 'workdays',
-    0x60: 'weekend',
-    0x01: 'mon',
-    0x02: 'tue',
-    0x04: 'wed',
-    0x08: 'thu',
-    0x10: 'fri',
-    0x20: 'sat',
-    0x40: 'sun',
-    0x55: 'mon-wed-fri-sun',
-    0x2a: 'tue-thu-sat',
-};
-
-
 const fzLocal = {
     aqara_s1_co2: {
         cluster: 'msCO2',
@@ -194,90 +178,6 @@ const fzLocal = {
                     break;
                 default:
                     meta.logger.warn(`zigbee-herdsman-converters:aqara_trv: Unknown key ${key} = ${value}`);
-                }
-            });
-            return result;
-        },
-    } satisfies Fz.Converter,
-    aqara_feeder: {
-        cluster: 'aqaraOpple',
-        type: ['attributeReport', 'readResponse'],
-        convert: (model, msg, publish, options, meta) => {
-            const result: KeyValue = {};
-            Object.entries(msg.data).forEach(([key, value]) => {
-                switch (parseInt(key)) {
-                case 0xfff1: {
-                    // @ts-expect-error
-                    const attr = value.slice(3, 7);
-                    // @ts-expect-error
-                    const len = value.slice(7, 8).readUInt8();
-                    // @ts-expect-error
-                    const val = value.slice(8, 8 + len);
-                    switch (attr.readInt32BE()) {
-                    case 0x04150055: // feeding
-                        result['feed'] = '';
-                        break;
-                    case 0x041502bc: { // feeding report
-                        const report = val.toString();
-                        result['feeding_source'] = {0: 'schedule', 1: 'manual', 2: 'remote'}[parseInt(report.slice(0, 2))];
-                        result['feeding_size'] = parseInt(report.slice(3, 4));
-                        break;
-                    }
-                    case 0x0d680055: // portions per day
-                        result['portions_per_day'] = val.readUInt16BE();
-                        break;
-                    case 0x0d690055: // weight per day
-                        result['weight_per_day'] = val.readUInt32BE();
-                        break;
-                    case 0x0d0b0055: // error ?
-                        result['error'] = utils.getFromLookup(val.readUInt8(), {1: true, 0: false});
-                        break;
-                    case 0x080008c8: { // schedule string
-                        const schlist = val.toString().split(',');
-                        const schedule: unknown[] = [];
-                        schlist.forEach((str: string) => { // 7f13000100
-                            const feedtime = Buffer.from(str, 'hex');
-                            schedule.push({
-                                'days': utils.getFromLookup(feedtime[0], daysLookup),
-                                'hour': feedtime[1],
-                                'minute': feedtime[2],
-                                'size': feedtime[3],
-                            });
-                        });
-                        result['schedule'] = schedule;
-                        break;
-                    }
-                    case 0x04170055: // indicator
-                        result['led_indicator'] = utils.getFromLookup(val.readUInt8(), {1: 'ON', 0: 'OFF'});
-                        break;
-                    case 0x04160055: // child lock
-                        result['child_lock'] = utils.getFromLookup(val.readUInt8(), {1: 'LOCK', 0: 'UNLOCK'});
-                        break;
-                    case 0x04180055: // mode
-                        result['mode'] = utils.getFromLookup(val.readUInt8(), {1: 'schedule', 0: 'manual'});
-                        break;
-                    case 0x0e5c0055: // serving size
-                        result['serving_size'] = val.readUInt8();
-                        break;
-                    case 0x0e5f0055: // portion weight
-                        result['portion_weight'] = val.readUInt8();
-                        break;
-                    case 0x080007d1: // ? 64
-                    case 0x0d090055: // ? 00
-                        meta.logger.warn(`zigbee-herdsman-converters:aqara_feeder: Unhandled attribute ${attr} = ${val}`);
-                        break;
-                    default:
-                        meta.logger.warn(`zigbee-herdsman-converters:aqara_feeder: Unknown attribute ${attr} = ${val}`);
-                    }
-                    break;
-                }
-                case 0x00ff: // 80:13:58:91:24:33:20:24:58:53:44:07:05:97:75:17
-                case 0x0007: // 00:00:00:00:1d:b5:a6:ed
-                case 0x00f7: // 05:21:14:00:0d:23:21:25:00:00:09:21:00:01
-                    meta.logger.debug(`zigbee-herdsman-converters:aqara_feeder: Unhandled key ${key} = ${value}`);
-                    break;
-                default:
-                    meta.logger.warn(`zigbee-herdsman-converters:aqara_feeder: Unknown key ${key} = ${value}`);
                 }
             });
             return result;
@@ -530,81 +430,6 @@ const tzLocal = {
             if (dict.hasOwnProperty(key)) {
                 await entity.read('aqaraOpple', [utils.getFromLookup(key, dict)], {manufacturerCode: 0x115F});
             }
-        },
-    } satisfies Tz.Converter,
-    aqara_feeder: {
-        key: ['feed', 'schedule', 'led_indicator', 'child_lock', 'mode', 'serving_size', 'portion_weight'],
-        convertSet: async (entity, key, value, meta) => {
-            const sendAttr = async (attrCode: number, value: number, length: number) => {
-                // @ts-expect-error
-                entity.sendSeq = ((entity.sendSeq || 0)+1) % 256;
-                // @ts-expect-error
-                const val = Buffer.from([0x00, 0x02, entity.sendSeq, 0, 0, 0, 0, 0]);
-                // @ts-expect-error
-                entity.sendSeq += 1;
-                val.writeInt32BE(attrCode, 3);
-                val.writeUInt8(length, 7);
-                let v = Buffer.alloc(length);
-                switch (length) {
-                case 1:
-                    v.writeUInt8(value);
-                    break;
-                case 2:
-                    v.writeUInt16BE(value);
-                    break;
-                case 4:
-                    v.writeUInt32BE(value);
-                    break;
-                default:
-                    // @ts-expect-error
-                    v = value;
-                }
-                await entity.write('aqaraOpple', {0xfff1: {value: Buffer.concat([val, v]), type: 0x41}},
-                    {manufacturerCode: 0x115f});
-            };
-            switch (key) {
-            case 'feed':
-                sendAttr(0x04150055, 1, 1);
-                break;
-            case 'schedule': {
-                const schedule: string[] = [];
-                // @ts-expect-error
-                value.forEach((item) => {
-                    const schedItem = Buffer.from([
-                        utils.getKey(daysLookup, item.days, 0x7f),
-                        item.hour,
-                        item.minute,
-                        item.size,
-                        0,
-                    ]);
-                    schedule.push(schedItem.toString('hex'));
-                });
-                const val = Buffer.concat([Buffer.from(schedule.join(',')), Buffer.from([0])]);
-                // @ts-expect-error
-                sendAttr(0x080008c8, val, val.length);
-                break;
-            }
-            case 'led_indicator':
-                sendAttr(0x04170055, utils.getFromLookup(value, {'OFF': 0, 'ON': 1}), 1);
-                break;
-            case 'child_lock':
-                sendAttr(0x04160055, utils.getFromLookup(value, {'UNLOCK': 0, 'LOCK': 1}), 1);
-                break;
-            case 'mode':
-                sendAttr(0x04180055, utils.getFromLookup(value, {'manual': 0, 'schedule': 1}), 1);
-                break;
-            case 'serving_size':
-                // @ts-expect-error
-                sendAttr(0x0e5c0055, value, 4);
-                break;
-            case 'portion_weight':
-                // @ts-expect-error
-                sendAttr(0x0e5f0055, value, 4);
-                break;
-            default: // Unknown key
-                meta.logger.warn(`zigbee-herdsman-converters:aqara_feeder: Unhandled key ${key}`);
-            }
-            return {state: {[key]: value}};
         },
     } satisfies Tz.Converter,
     aqara_fp1_region_upsert: {
@@ -3424,8 +3249,8 @@ const definitions: Definition[] = [
         model: 'ZNCWWSQ01LM',
         vendor: 'Xiaomi',
         description: 'Aqara pet feeder C1',
-        fromZigbee: [fzLocal.aqara_feeder],
-        toZigbee: [tzLocal.aqara_feeder],
+        fromZigbee: [xiaomi.fromZigbee.aqara_feeder],
+        toZigbee: [xiaomi.toZigbee.aqara_feeder],
         exposes: [
             e.enum('feed', ea.STATE_SET, ['', 'START']).withDescription('Start feeding'),
             e.enum('feeding_source', ea.STATE, ['schedule', 'manual', 'remote']).withDescription('Feeding source'),
