+++ test/xiaomi.test.js
@@ -1,4 +1,4 @@
-const {trv} = require('../src/lib/xiaomi');
+const {trv, fromZigbee} = require('../src/lib/xiaomi');
 
 describe('lib/xiaomi', () => {
     describe('trv', () => {
@@ -243,6 +243,20 @@ describe('lib/xiaomi', () => {
                 const schedule = trv.parseSchedule('');
 
                 expect(schedule).toEqual({days: [], events: []});
+            });
+        });
+
+        describe('Feeder schedule', () => {
+            it('Schedule 0 days', () => {
+                const data = Buffer.from([0,5,43,8,0,8,200,2,47,47]);
+                const result = fromZigbee.aqara_feeder.convert(null, {data: {'65521': data}}, null, null, {logger: {warn: jest.fn(), debug: jest.fn()}});
+                expect(result).toStrictEqual({ schedule: [] });
+            });
+
+            it('Schedule 1 day', () => {
+                const data = Buffer.from([0,5,9,8,0,8,200,10,55,70,48,49,48,49,48,49,48,48]);
+                const result = fromZigbee.aqara_feeder.convert(null, {data: {'65521': data}}, null, null, {logger: {warn: jest.fn(), debug: jest.fn()}});
+                expect(result).toStrictEqual({ schedule: [ { days: 'everyday', hour: 1, minute: 1, size: 1 } ] });
             });
         });
     });
